プロフィール画像用カラム

+++ db/migrate/20200618112127_devise_create_users.rb
@@ -34,7 +34,7 @@ class DeviseCreateUsers < ActiveRecord::Migration[5.2]
 
       t.string :name, null: false, default: ""
       t.text :introduction
-      t.bigint :profile_image_id
+      t.integer :profile_image_id
 
       t.timestamps null: false
     end
